+++ riverdriver/riverdrivertest/driver_client_test.go
@@ -516,7 +516,10 @@ func ExerciseClient[TTx any](ctx context.Context, t *testing.T,
 
 		client, bundle := setup(t)
 
-		job := testfactory.Job(ctx, t, bundle.exec, &testfactory.JobOpts{Schema: bundle.schema})
+		job := testfactory.Job(ctx, t, bundle.exec, &testfactory.JobOpts{
+			Schema: bundle.schema,
+			Tags:   []string{"all-args-tag"},
+		})
 
 		listRes, err := client.JobList(ctx,
 			river.NewJobListParams().
@@ -524,7 +527,8 @@ func ExerciseClient[TTx any](ctx context.Context, t *testing.T,
 				Kinds(job.Kind).
 				Priorities(int16(min(job.Priority, math.MaxInt16))). //nolint:gosec
 				Queues(job.Queue).
-				States(job.State),
+				States(job.State).
+				Tags("ALL-ARGS-TAG"),
 		)
 		require.NoError(t, err)
 		require.Len(t, listRes.Jobs, 1)
@@ -550,6 +554,24 @@ func ExerciseClient[TTx any](ctx context.Context, t *testing.T,
 		require.NoError(t, err)
 		require.Len(t, listRes.Jobs, 1)
 		require.Equal(t, job.ID, listRes.Jobs[0].ID)
+	})
+
+	t.Run("JobListTags", func(t *testing.T) {
+		t.Parallel()
+
+		client, bundle := setup(t)
+
+		var (
+			job1 = testfactory.Job(ctx, t, bundle.exec, &testfactory.JobOpts{Schema: bundle.schema, Tags: []string{"alpha", "shared"}})
+			job2 = testfactory.Job(ctx, t, bundle.exec, &testfactory.JobOpts{Schema: bundle.schema, Tags: []string{"beta"}})
+			_    = testfactory.Job(ctx, t, bundle.exec, &testfactory.JobOpts{Schema: bundle.schema, Tags: []string{"gamma"}})
+		)
+
+		listRes, err := client.JobList(ctx, river.NewJobListParams().Tags("ALPHA", "BETA"))
+		require.NoError(t, err)
+		require.Len(t, listRes.Jobs, 2)
+		require.Equal(t, job1.ID, listRes.Jobs[0].ID)
+		require.Equal(t, job2.ID, listRes.Jobs[1].ID)
 	})
 
 	t.Run("JobListTx", func(t *testing.T) {
@@ -578,7 +600,10 @@ func ExerciseClient[TTx any](ctx context.Context, t *testing.T,
 
 		tx, execTx := beginTx(ctx, t, bundle)
 
-		job := testfactory.Job(ctx, t, execTx, &testfactory.JobOpts{Schema: bundle.schema})
+		job := testfactory.Job(ctx, t, execTx, &testfactory.JobOpts{
+			Schema: bundle.schema,
+			Tags:   []string{"all-args-tag"},
+		})
 
 		listRes, err := client.JobListTx(ctx, tx,
 			river.NewJobListParams().
@@ -586,7 +611,8 @@ func ExerciseClient[TTx any](ctx context.Context, t *testing.T,
 				Kinds(job.Kind).
 				Priorities(int16(min(job.Priority, math.MaxInt16))). //nolint:gosec
 				Queues(job.Queue).
-				States(job.State),
+				States(job.State).
+				Tags("ALL-ARGS-TAG"),
 		)
 		require.NoError(t, err)
 		require.Len(t, listRes.Jobs, 1)